[app][github] Allow specify repo to list labels
- previously it was hard coded to use at15/papers-i-read
  - [ ] TODO: it might be better merge the logic with the git app on
handling repository urls, might it use for GitHub enterprise
- also split sub commands to their own files, i.e. web/static.go,
git/clone.go
